fix: harden audio marker slicing
